refactor: Move HangfireJobRegistrationHelper from Application to API layer
Relocates HangfireJobRegistrationHelper.cs from
Anela.Heblo.Application/Features/BackgroundJobs/Services/ to
Anela.Heblo.API/Infrastructure/Hangfire/ so the Application layer no
longer takes a compile-time dependency on the Hangfire static API
through this helper. Public surface, method signatures, reflection
dispatch, and runtime behavior are byte-for-byte unchanged.

- Source helper moved with git mv; namespace updated to
  Anela.Heblo.API.Infrastructure.Hangfire (matches sibling adapters).
- Tests moved to backend/test/Anela.Heblo.Tests/Infrastructure/Hangfire/
  mirroring the new production path; using directive updated.
- RecurringJobDiscoveryService.cs: stale using directive removed.
- HangfireRecurringJobScheduler.cs: using retained — that namespace
  still provides IHangfireRecurringJobScheduler.

Hangfire.Core PackageReference on Anela.Heblo.Application.csproj is
retained because FailedJobsTile, DashboardModule (JobStorage.Current),
GenerateArticleJob, GenerateArticleHandler, ProductExportDownloadJob,
PlaudPollingJob, and IHangfireJobEnqueuer still depend on it. Removing
those references is tracked separately (spec Out of Scope; arch-review
Amendment B).
